changed return type for AddItems
git-svn-id: http://svn.xp-dev.com/svn/wxextension@1016 f22100f3-aa73-48fe-a4fb-fd497bb32605

+++ trunk/extension/include/wx/extension/report/listview.h
@@ -137,7 +137,7 @@ protected:
   void OnList(wxListEvent& event);
   void OnMouse(wxMouseEvent& event);
 private:
-  int AddItems();
+  size_t AddItems();
   void DeleteDoubles();
   const wxString GetFindInCaption(int id); // cannot be const
   void Initialize(const wxExLexer* lexer);
